Use a ShadowPicker for Resources.
