+++ robolectric/src/test/java/org/robolectric/AttributeSetBuilderTest.java
@@ -339,6 +339,7 @@ public class AttributeSetBuilderTest {
       Robolectric.buildAttributeSet()
             .setStyleAttribute("@style/non_existent_style")
             .build();
+      fail();
     } catch (IllegalArgumentException e) {
       assertThat(e)
           .hasMessageThat()
