+++ shadows/framework/src/main/java/org/robolectric/shadows/ShadowArscAssetManager.java
@@ -821,14 +821,14 @@ public class ShadowArscAssetManager extends ShadowAssetManager {
     return block;
   }
 
-  /*package*/ static final int STYLE_NUM_ENTRIES = 6;
-  /*package*/ static final int STYLE_TYPE = 0;
-  /*package*/ static final int STYLE_DATA = 1;
-  /*package*/ static final int STYLE_ASSET_COOKIE = 2;
-  /*package*/ static final int STYLE_RESOURCE_ID = 3;
-
-  /* Offset within typed data array for native changingConfigurations. */
-  static final int STYLE_CHANGING_CONFIGURATIONS = 4;
+  // /*package*/ static final int STYLE_NUM_ENTRIES = 6;
+  // /*package*/ static final int STYLE_TYPE = 0;
+  // /*package*/ static final int STYLE_DATA = 1;
+  // /*package*/ static final int STYLE_ASSET_COOKIE = 2;
+  // /*package*/ static final int STYLE_RESOURCE_ID = 3;
+  //
+  // /* Offset within typed data array for native changingConfigurations. */
+  // static final int STYLE_CHANGING_CONFIGURATIONS = 4;
 
   /*package*/ static final int STYLE_DENSITY = 5;
